# refactor: refine validation for decimal128 array