@liukun4515 2022-08-12
Benchmark result:
before this refactor:
```
validate_decimal128_array 20000
                        time:   [56.566 us 56.867 us 57.195 us]
                        change: [-0.7933% -0.1263% +0.5894%] (p = 0.72 > 0.05)
                        No change in performance detected.
Found 7 outliers among 100 measurements (7.00%)
  6 (6.00%) high mild
  1 (1.00%) high severe
```

After this refactor:

```
validate_decimal128_array 20000
                        time:   [34.534 us 34.711 us 34.895 us]
                        change: [-39.461% -39.030% -38.571%] (p = 0.00 < 0.05)
                        Performance has improved.
```

The performance improves about 40%